@@ -1,8 +1,8 @@
 source 'http://rubygems.org'
 
-gem 'rails',          '>=3.1.0'
-gem 'rails_autolink', '>=1.0.4'
-gem 'jquery-rails',   '>=1.0.0'
+gem 'rails',          '~> 4.0.0'
+gem 'rails_autolink', '~> 1.1.4'
+gem 'jquery-rails',   '~> 3.0.4'
 
 # gem 'will_paginate',  '>=3.0.2'
 # gem 'kaminari',       '>=0.0.0'
@@ -10,4 +10,5 @@ gem 'jquery-rails',   '>=1.0.0'
 group :test do
   gem 'sqlite3'
   gem 'jeweler'
+  gem "pry-rails"
 end
