@@ -1,40 +1,40 @@
 class Admin::Blog::TagsController < Admin::Blog::BaseController
-  
+
   before_filter :build_tag, :only => [:new, :create]
   before_filter :load_tag,  :only => [:edit, :update, :destroy]
 
   def index
     @tags = Blog::Tag.order('is_category DESC', :name)
   end
-  
+
   def edit
     render
   end
-  
+
   def new
     render
   end
-  
+
   def update
-    @tag.update_attributes!(params[:tag])
+    @tag.update_attributes!(tag_params)
     flash[:notice] = 'Blog Tag updated'
     redirect_to :action => :index
-    
+
   rescue ActiveRecord::RecordInvalid
     flash.now[:error] = 'Failed to update Blog Tag'
     render :action => :edit
   end
-  
+
   def create
     @tag.save!
     flash[:notice] = 'Blog Tag created'
     redirect_to :action => :index
-    
+
   rescue ActiveRecord::RecordInvalid
     flash.now[:error] = 'Failed to create Blog Tag'
     render :action => :new
   end
-  
+
   def destroy
     @tag.destroy
     flash[:notice] = 'Blog Tag removed'
@@ -42,15 +42,22 @@ class Admin::Blog::TagsController < Admin::Blog::BaseController
   end
 
 protected
-  
+
   def build_tag
-    @tag = Blog::Tag.new(params[:tag])
+    @tag = Blog::Tag.new(tag_params)
   end
-  
+
   def load_tag
     @tag = Blog::Tag.find(params[:id])
   rescue ActiveRecord::RecordNotFound
     flash[:error] = 'Blog Tag not found'
-    redirect_to :action => :index  
+    redirect_to :action => :index
   end
+
+private
+
+  def tag_params
+    params.fetch(:tag, {}).permit(:name, :is_category)
+  end
+
 end
