@@ -1,7 +1,7 @@
 require File.expand_path('../../test_helper', File.dirname(__FILE__))
 
 class Blog::PostsControllerTest < ActionController::TestCase
-  
+
   def test_get_index
     get :index
     assert_response :success
@@ -9,7 +9,7 @@ class Blog::PostsControllerTest < ActionController::TestCase
     assert assigns(:posts)
     assert_equal 1, assigns(:posts).size
   end
-  
+
   def test_get_index_as_rss
     get :index, :format => :rss
     assert_response :success
@@ -17,14 +17,14 @@ class Blog::PostsControllerTest < ActionController::TestCase
     assert assigns(:posts)
     assert_equal 1, assigns(:posts).size
   end
-  
+
   def test_get_index_with_unpublished
     blog_posts(:default).update_attribute(:is_published, false)
     get :index
     assert_response :success
     assert_equal 0, assigns(:posts).size
   end
-  
+
   def test_get_index_with_pagination
     if defined?(WillPaginate) || defined?(Kaminari)
       get :index, :page => 99
@@ -32,61 +32,61 @@ class Blog::PostsControllerTest < ActionController::TestCase
       assert_equal 0, assigns(:posts).size
     end
   end
-  
+
   def test_get_index_for_tagged
     get :index, :tag => blog_tags(:tag).name
     assert_response :success
     assert_equal 1, assigns(:posts).size
-    
+
     get :index, :tag => 'invalid'
     assert_response :success
     assert_equal 0, assigns(:posts).size
   end
-  
+
   def test_get_index_for_categorized
     get :index, :category => blog_tags(:category).name
     assert_response :success
     assert_equal 1, assigns(:posts).size
-    
+
     get :index, :category => 'invalid'
     assert_response :success
     assert_equal 0, assigns(:posts).size
   end
-  
+
   def test_get_index_for_year_archive
     get :index, :year => 2012
     assert_response :success
     assert_equal 1, assigns(:posts).size
-    
+
     get :index, :year => 1999
     assert_response :success
     assert_equal 0, assigns(:posts).size
   end
-  
+
   def test_get_index_for_month_archive
     get :index, :year => 2012, :month => 1
     assert_response :success
     assert_equal 1, assigns(:posts).size
-    
+
     get :index, :year => 2012, :month => 12
     assert_response :success
     assert_equal 0, assigns(:posts).size
   end
-  
+
   def test_get_show
     get :show, :id => blog_posts(:default)
     assert_response :success
     assert_template :show
     assert assigns(:post)
   end
-  
+
   def test_get_show_unpublished
     post = blog_posts(:default)
     post.update_attribute(:is_published, false)
     get :show, :id => post
     assert_response 404
   end
-  
+
   def test_get_show_with_date
     post = blog_posts(:default)
     get :show, :year => post.year, :month => post.month, :slug => blog_posts(:default).slug
@@ -94,17 +94,17 @@ class Blog::PostsControllerTest < ActionController::TestCase
     assert_template :show
     assert assigns(:post)
   end
-  
+
   def test_get_show_with_date_invalid
     get :show, :year => '1999', :month => '99', :slug => 'invalid'
     assert_response 404
   end
-  
+
   def test_get_show_with_disqus
     ComfyBlog.config.disqus_shortname = 'test'
     post = blog_posts(:default)
     get :show, :year => post.year, :month => post.month, :slug => blog_posts(:default).slug
     assert_response :success
   end
-  
+
 end
